First attempt at VGA - not fast but it worked! VGA 640x480x4 is currently the only implemented configuration but more could be added.

+++ Kernel/Libraries/Kernel.IO/IOPort.cs
@@ -157,7 +157,7 @@ namespace Kernel.IO
         /// <returns>The value read.</returns>
         [NoDebug]
         [NoGC]
-        public ushort Read_Byte(ushort offset)
+        public byte Read_Byte(ushort offset)
         {
             return doRead_Byte((ushort)(Port + offset));
         }
